fix: move check expiring command to movipass
